Fix darwin and windows resource usage collection
